fix(update): load aggregates without exceptions
Introduce explicitly fallible aggregate ownership, ignore seek outcomes, use the low 32 bits of the reported extent, and apply the specified wrapping size and controller-offset arithmetic. Allocation and short-read failures now emit their exact diagnostics.\n\nAll 803 host tests pass and the ESP-IDF live target build succeeds.

+++ components/application/src/update/update_application.cpp
@@ -62,8 +62,8 @@ void UpdateApplicationService::stage_controller(
     if (package.header.controller_size == 0U) {
         return;
     }
-    const std::size_t controller_offset =
-        core::update_package_header_size + package.header.mainboard_size;
+    const std::uint32_t controller_offset = static_cast<std::uint32_t>(
+        core::update_package_header_size + package.header.mainboard_size);
     const core::BytesView controller_image(
         package.bytes.data() + controller_offset,
         package.header.controller_size);
